move my tests to separate file
Testing function became private so I had to move these tests to package internal. Moving all tests to this package is very hard because it creates cyclic dependencies with package testtools and almost all project structure has to be changed
