+++ internal/backup_test.go
@@ -7,9 +7,6 @@ import (
 	"github.com/wal-g/wal-g/internal"
 	"github.com/wal-g/wal-g/testtools"
 	"github.com/wal-g/wal-g/utility"
-	"io/ioutil"
-	"os"
-	"path/filepath"
 	"testing"
 )
 
@@ -137,98 +134,4 @@ func TestFetchSentinelReturnErrorWhenSentinelUnmarshallable(t *testing.T) {
 
 	assert.Error(t, err)
 	assert.Equal(t, errorMessage, err.Error()[:len(errorMessage)])
-}
-
-func createTempDir(prefix string) (name string, err error) {
-	cwd, err := filepath.Abs("./")
-	if err != nil {
-		return "", err
-	}
-
-	dir, err := ioutil.TempDir(cwd, prefix)
-	if err != nil {
-		return "", err
-	}
-
-	return dir, nil
-}
-
-func TestIsDirectoryEmpty_ReturnsTrue_WhenDirectoryIsEmpty(t *testing.T) {
-	dir, err := createTempDir("empty")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.Remove(dir)
-
-	actual, err := internal.IsDirectoryEmpty(dir)
-
-	assert.True(t, actual)
-}
-
-func TestIsDirectoryEmpty_ReturnsFalse_WhenOneFileIsInDirectory(t *testing.T) {
-	dir, err := createTempDir("not_empty")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.Remove(dir)
-
-	file, err := ioutil.TempFile(dir, "file")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.Remove(file.Name())
-
-	actual, err := internal.IsDirectoryEmpty(dir)
-
-	assert.False(t, actual)
-}
-
-func TestIsDirectoryEmpty_ReturnsFalse_WhenSeveralFilesAreInDirectory(t *testing.T) {
-	dir, err := createTempDir("not_empty")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.Remove(dir)
-
-	for i := 0; i < 3; i++ {
-		file, err := ioutil.TempFile(dir, "file")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.Remove(file.Name())
-	}
-
-	actual, err := internal.IsDirectoryEmpty(dir)
-
-	assert.False(t, actual)
-}
-
-func TestIsDirectoryEmpty_ReturnsFalse_WhenNestedDirectoryIsInDirectory(t *testing.T) {
-	dir, err := createTempDir("not_empty")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.Remove(dir)
-
-	nested, err := ioutil.TempDir(dir, "nested")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.Remove(nested)
-
-	actual, err := internal.IsDirectoryEmpty(dir)
-
-	assert.False(t, actual)
-}
-
-func TestIsDirectoryEmpty_ReturnsTrue_WhenDirectoryDoesntExist(t *testing.T) {
-	dir, err := createTempDir("not_existing")
-	if err != nil {
-		t.Fatal(err)
-	}
-	os.Remove(dir)
-
-	actual, err := internal.IsDirectoryEmpty(dir)
-
-	assert.True(t, actual)
 }
